tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/context.unrecognised.badsuffix.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: context.unrecognised.badsuffix</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>context.unrecognised.badsuffix</h1>
<p class="desc">Context name "2d" plus a suffix is unrecognised</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Context name \"2d\" plus a suffix is unrecognised");
_addTest(function(canvas, ctx) {

_assertSame(canvas.getContext("2d#"), null, "canvas.getContext(\"2d#\")", "null");


});
</script>
